modify: move DndContext from TabList to WindowContainer
